ring: Consistently use size_t to specify ring size
The mm_len member of struct ring is of type size_t, but in the code
paths leading to set it, unsigned int is used. In circumstances where
unsigned int is 32 bit and size_t is 64 bit, this could lead to an
integer overflow, which causes an improper ring size being mmap()'ed in
mmap_ring_generic().

In order to prevent this, consistently use size_t to store the ring
size, since this is also what mmap() takes as its `length' parameter.

This now allows to specify ring sizes larger than 4 GiB for both
netsniff-ng and trafgen (fixes #90).

Reported-by: Jon Schipp <jonschipp@gmail.com>
Reported-by: Michał Purzyński <michalpurzynski1@gmail.com>

+++ ring.h
@@ -67,7 +67,7 @@ static inline void next_rnd_slot(unsigned int *it, struct ring *ring)
 	*it = rand() % ring->layout.tp_frame_nr;
 }
 
-static inline unsigned int ring_size(char *ifname, unsigned int size)
+static inline size_t ring_size(char *ifname, size_t size)
 {
 	if (size > 0)
 		return size;
